Make trend charts theme-aware.
Add a `JenkinsPalette` that can be used to obtain colors that are
supported by the current Jenkins theme. All line charts can now handle
fixed colors or CSS color variables that will be resolved on the client
side using the actual values of the selected theme.
