+++ src/main/resources/io/jenkins/plugins/echarts.jelly
@@ -13,13 +13,8 @@ Use it like <st:adjunct includes="io.jenkins.plugins.echarts"/>
 
   <link type="text/css" rel="stylesheet" href="${resURL}/plugin/echarts-api/css/jenkins-style.css"/>
 
-  <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/echarts-scope.js"/>
-  <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/chart-configuration.js"/>
-  <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/trend-chart.js"/>
-  <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/zoomable-trend-chart.js"/>
-  <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/configurable-trend-chart.js"/>
+  <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/echarts-api.js"/>
   <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/pie-chart.js"/>
   <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/progress-chart.js"/>
-  <script type="text/javascript" src="${resURL}/plugin/echarts-api/js/escapeMetaCharacters.js"/>
 
 </j:jelly>
