--- src/main/webapp/js/echarts-scope.js
@@ -1,3 +1,0 @@
-/* exported echartsJenkinsApi */
-function EChartsJenkinsApi () { /* Empty placeholder */ }
-const echartsJenkinsApi = new EChartsJenkinsApi();
