--- src/main/webapp/js/escapeMetaCharacters.js
@@ -1,3 +1,0 @@
-function escapeMetaCharacters(string) {
-    return string.replaceAll(/[.*+?^${}()|[\]\\]/g, '\\$&')
-}
