+++ src/main/webapp/js/pie-chart.js
@@ -5,7 +5,7 @@ jQuery3(document).ready(function () {
      */
     function renderPieCharts() {
         /**
-         * Renders a trend chart in the a div using ECharts.
+         * Renders a trend chart in a div using ECharts.
          *
          * @param {String} chartDivId - the ID of the div where the chart should be shown in
          */
@@ -35,7 +35,7 @@ jQuery3(document).ready(function () {
                 }
             }
 
-            const chartPlaceHolder = jQuery3("#" + escapeMetaCharacters(chartDivId));
+            const chartPlaceHolder = jQuery3("#" + echartsJenkinsApi.escapeMetaCharacters(chartDivId));
             const model = JSON.parse(chartPlaceHolder.attr('data-chart-model'));
             const title = chartPlaceHolder.attr('data-title');
             const chartDiv = chartPlaceHolder[0];
