+++ src/main/webapp/js/progress-chart.js
@@ -5,12 +5,12 @@ jQuery3(document).ready(function () {
      */
     function renderProgressChart() {
         /**
-         * Renders a trend chart in the a div using ECharts.
+         * Renders a trend chart in a div using ECharts.
          *
          * @param {String} chartDivId - the ID of the div where the chart should be shown in
          */
         function renderProgressChart(chartDivId) {
-            const chartPlaceHolder = jQuery3("#" + escapeMetaCharacters(chartDivId));
+            const chartPlaceHolder = jQuery3("#" + echartsJenkinsApi.escapeMetaCharacters(chartDivId));
             const model = JSON.parse(chartPlaceHolder.attr('data-chart-model'));
             const title = chartPlaceHolder.attr('data-title');
             const tooltip = chartPlaceHolder.attr('data-tooltip');
